bridge/requirements.txt
pyserial>=3.5
websockets>=13.0
